+++ lib/components/touch_scale_context.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/animation.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter_touch_scale/flutter_touch_scale.dart';
 
 @protected
 mixin TouchScaleContext {
@@ -9,4 +10,5 @@ mixin TouchScaleContext {
   Duration get reverseDuration;
   Curve get reverseCurve;
   Duration? get previewDuration;
+  TouchScaleCallPhase get callPhase;
 }